Add unit test for binary writer
